+++ lib/resource_policy/rails.rb
@@ -2,4 +2,5 @@
 
 # Incudes resource policy and rails specific helpers
 require 'resource_policy'
-require 'resource_policy/validators/action_policy_validator'
+require 'resource_policy/validators/action_validator'
+require 'resource_policy/validators/attributes_validator'