+++ CHANGELOG.md
@@ -5,8 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.0.0]
 
+* Updated: Ruby version updated to 4.0
 * Added: short-hand method for getting allowed actions
 * Added: support for proc type validator options
 * Added: nested read protection. `protected_resource` now returns a protected resource for